requirements-edge.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
tqdm==4.66.1
